refactored file loaders into utils.py
